👌 split simpleTestDoubles
