+++ packages/core/test/emulate/cookie.ts
@@ -1,0 +1,13 @@
+export function stubCookie() {
+  let cookie = ''
+  return {
+    getSpy: spyOnProperty(Document.prototype, 'cookie', 'get').and.callFake(() => cookie),
+    setSpy: spyOnProperty(Document.prototype, 'cookie', 'set').and.callFake((newCookie) => {
+      cookie = newCookie
+    }),
+    currentValue: () => cookie,
+    setCurrentValue: (newCookie: string) => {
+      cookie = newCookie
+    },
+  }
+}
