+++ packages/core/test/emulate/visibilityState.ts
@@ -1,0 +1,12 @@
+export function setPageVisibility(visibility: 'visible' | 'hidden') {
+  Object.defineProperty(document, 'visibilityState', {
+    get() {
+      return visibility
+    },
+    configurable: true,
+  })
+}
+
+export function restorePageVisibility() {
+  delete (document as any).visibilityState
+}
