+++ packages/core/test/index.ts
@@ -1,11 +1,17 @@
 export * from './browserChecks'
 export * from './buildEnv'
 export * from './collectAsyncCalls'
+export * from './requests'
 export * from './emulate/createNewEvent'
 export * from './emulate/location'
 export * from './emulate/mockClock'
-export * from './requests'
 export * from './emulate/stubReportApis'
 export * from './emulate/stubZoneJs'
 export * from './emulate/syntheticsWorkerValues'
-export * from './emulate/simpleTestDoubles'
+export * from './emulate/visibilityState'
+export * from './emulate/navigatorOnLine'
+export * from './emulate/navigatorOnLine'
+export * from './emulate/eventBridge'
+export * from './emulate/eventBridge'
+export * from './emulate/windowOnError'
+export * from './emulate/cookie'
